Add plot_functions function to kinda_tidy.py
plot_functions takes a string for the independent variable name, a range for this variable, optionally a number of points in the domain on which to evaluate the functions, and a sequence of functions of the form name=lambda df: compute_value(df.varname) and produces a ggplot object plotting the functions on the range.  The ggplot object can be further manipulated e.g. change axes to log scale, to create a final plot.